requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.3
dataclasses-json==0.5.13
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
faiss-cpu==1.7.4
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-aiplatform==1.28.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-discoveryengine==0.9.1
google-cloud-resource-manager==1.10.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.56.2
grpcio-status==1.56.2
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.240
langsmith==0.0.14
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.25.1
openapi-schema-pydantic==1.2.4
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
Pygments==2.15.1
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data==0.6.2
tenacity==8.2.2
tornado==6.3.2
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6
yarl==1.9.2
